==> first_submission/4. pca_on_bert_embeddings.ipynb <==
import pandas as pd

# Load dataset
df = pd.read_csv("final_test_bert_data.csv")
old_df=df

df

# cols_to_keep = list(range(1, 769))

# # Select the columns
# df = df.iloc[:, cols_to_keep]
# df


# Drop unwanted columns
drop_cols = ['Value', 'image_link', 'sample_id', 'Unit_grams', 'Unit_ml']
X = df.drop(columns=drop_cols + ['price'], errors='ignore')  # drop price too from features
X


from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA


scaler = StandardScaler()

X_scaled = scaler.fit_transform(X)

# Apply PCA to select top 300 components
n_components = 300
pca = PCA(n_components=n_components)
X_pca = pca.fit_transform(X_scaled)

# Create DataFrame with PCA features
pca_columns = [f'PC_{i+1}' for i in range(n_components)]
pca_df = pd.DataFrame(X_pca, columns=pca_columns)

pca_df

pca_df['sample_id'] = df['sample_id']
pca_df


# Save to CSV
pca_df.to_csv('pca_300_test_bert_features.csv', index=False)

print("CSV saved: pca_300_features.csv")
print(f"Explained variance by top 300 components: {sum(pca.explained_variance_ratio_):.4f}")
print(pca_df.head())
