==> first_submission/6. merge_pca_vit_with_pca_bert.ipynb <==
import pandas as pd
df_vit = pd.read_csv("pca_300_test_vit_features.csv")
df_vit.tail()



# Remove duplicate rows based on 'sample_id'
df_vit = df_vit.drop_duplicates(subset='sample_id', keep='first')

df_bert = pd.read_csv("pca_300_test_bert_features.csv")

df_bert

df_vit


merged_df = pd.merge(df_bert, df_vit, on='sample_id', how='inner')

sample=merged_df['sample_id']

merged=merged_df.drop(columns='sample_id')

merged['sample_id']=sample
merged

merged.to_csv('merged_pca_test.csv',index=False)

import pandas as pd
priced_data=pd.read_csv('final_test_bert_data.csv')

merged_data=pd.read_csv('merged_pca_test.csv')

# Your list of columns to add
insert_cols = ['Value', 'image_link', 'Unit_grams', 'Unit_ml']

# --- Solution ---
merged_data[insert_cols] = priced_data[insert_cols]

merged_data

merged_data.to_csv('final_pca_whole_test_data.csv',index=False)

merged_data
